% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmmeans-package.R
\docType{package}
\name{kmmeans-package}
\alias{kmmeans-package}
\alias{_PACKAGE}
\title{kmmeans: K-means for observations with missing data}
\description{
K-means for observations with missing data.
}
\details{
The package kmmeans provides the function [kmmeans()], which performs
k-means clustering on data with missing values.
}
\author{
\strong{Maintainer}: Andrew Lithio \email{lithio...}

}
\keyword{internal}
