% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kmmeans.R
\name{kmmeans}
\alias{kmmeans}
\title{K-means clustering with missing data}
\usage{
kmmeans(data, K, n.init = 1)
}
\arguments{
\item{data}{A data frame of data with possible missing values (NA).}

\item{K}{Number of clusters to fit.}

\item{n.init}{Number of k-means++ initializations.}
}
\value{
A list with k-means clustering solution.
\itemize{
\item criterion - The minimum realized objective function value.
\item cluster.criteria - A vector of the minimum realized objective per
cluster.
\item cluster.sizes - The sizes of the clusters in the best solution.
\item partition - The partition of observations into clusters in the best
solution.
\item modes - The estimated modes of each cluster in K x p integer matrix.
\item average.criterion - The average objective function achieved across
multiple initializations.
\item number.initializations - The number of initializations performed.
\item best.ari - If true clusters provided, then the best achieved ARI,
 perhaps not from the best solution, as measured by the objective function.
\item average.ari - If true clusters provided, then the average ARI achieved
across all initializations.
}
}
\description{
Estimate means and partition of numerical data with missing values.
}
\examples{

}
\references{

}
